+++ mobile/android/android-components/components/feature/awesomebar/src/main/java/mozilla/components/feature/awesomebar/AwesomeBarFeature.kt
@@ -52,6 +52,7 @@ class AwesomeBarFeature(
         ))
 
         awesomeBar.setOnStopListener { toolbar.displayMode() }
+        awesomeBar.setOnEditSuggestionListener(toolbar::setSearchTerms)
     }
 
     /**